Se solucionó lo del depósito, y ahora es actualizado

+++ src/logic/Responsable.java
@@ -3,10 +3,10 @@ package logic;
 public class Responsable extends Persona{
 
     private String direccion;
-    private int celular;
-    private int cedula;
+    private long celular;
+    private long cedula;
 
-    public Responsable(String nombre, String direccion, int celular, int cedula) {
+    public Responsable(String nombre, String direccion, long celular, long cedula) {
         super(nombre);
         this.direccion = direccion;
         this.celular = celular;
@@ -21,19 +21,19 @@ public class Responsable extends Persona{
         this.direccion = direccion;
     }
 
-    public int getCedula() {
+    public long getCedula() {
         return cedula;
     }
 
-    public void setCedula(int cedula) {
+    public void setCedula(long cedula) {
         this.cedula = cedula;
     }
 
-    public int getCelular() {
+    public long getCelular() {
         return celular;
     }
 
-    public void setCelular(int celular) {
+    public void setCelular(long celular) {
         this.celular = celular;
     }
 }
